Geometry Data contained within a class
I should probably move things like setting the VAO, VBO and such into the geometry class.

+++ include/shader.hpp
@@ -160,7 +160,7 @@ namespace Xplor
 		// Program ID
 		uint32_t m_shaderID{};
 
-		uint16_t compileShader(int shaderType, const char * shaderSource) const;
+		GLuint compileShader(int shaderType, const char * shaderSource) const;
 
 	private:
 		std::string m_vertexPath{};
